# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Talertech POS API",
    "version": "1.0.0",
...
    "contact": { "name": "Talertech POS" }
  },
  "servers": [
    { "url": "https://pos-starter-kit.vercel.app", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token returned in the `set-auth-token` header on sign-in. Used by the Flutter mobile client on every request; web clients may instead rely on the httpOnly session cookie."
      },
      "cronSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "Value of the `CRON_SECRET` env var. Only enforced when NODE_ENV=production; intended for Vercel Cron only, not app clients."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "error": { "type": "string" }
        },
        "required": ["success", "message"]
      },
...
    },
...
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new account",
        "description": "Explicit wrapper around better-auth's sign-up/email endpoint. Creates the user (`plan=free`, trial not started until onboarding) and, on success, signs them in immediately — the response forwards the session cookie (web) and the `set-auth-token` bearer token (mobile) exactly like `/api/auth/login` does.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["name", "email", "password"],
            "properties": {
              "name": { "type": "string" },
              "email": { "type": "string", "format": "email" },
              "password": { "type": "string", "format": "password", "minLength": 8 },
              "image": { "type": "string", "description": "Optional avatar URL." }
            }
          } } }
        },
        "responses": {
          "200": {
            "description": "Registered and signed in. Bearer token is in the `set-auth-token` response header.",
            "headers": { "set-auth-token": { "schema": { "type": "string" }, "description": "Use as `Authorization: Bearer <token>` on subsequent requests." } },
            "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "object", "properties": { "token": { "type": "string" }, "user": { "type": "object" } } } } } } }
          },
          "400": { "description": "Missing fields, or email already registered.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Sign in with email and password",
        "description": "Explicit wrapper around better-auth's sign-in/email endpoint. On success, forwards the session cookie (web) and the `set-auth-token` bearer token (mobile) — send that value back as `Authorization: Bearer <token>` on every subsequent request. This is the endpoint the Flutter client should call.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "required": ["email", "password"],
            "properties": {
              "email": { "type": "string", "format": "email" },
              "password": { "type": "string", "format": "password" },
              "rememberMe": { "type": "boolean", "default": true }
            }
          } } }
        },
        "responses": {
          "200": {
            "description": "Signed in.",
            "headers": { "set-auth-token": { "schema": { "type": "string" }, "description": "Use as `Authorization: Bearer <token>` on subsequent requests." } },
            "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "object", "properties": { "token": { "type": "string" }, "user": { "type": "object" } } } } } } }
          },
          "400": { "description": "Missing email/password.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Invalid email or password.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}